requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-cleanup==9.0.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-redis==5.4.0
et-xmlfile==1.1.0
Faker==30.8.2
idna==3.10
kombu==5.4.0
lml==0.1.0
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
texttable==1.7.0
transaction==5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.0
xlwings==0.33.3
zope.interface==7.1.1
